player-refactoring: BaseCharacter
