+++ Source/MyShooterThemUp/Private/Player/STUPlayerCharacter.cpp
@@ -1,0 +1,82 @@
+// Shoot Them Up Game, All Rights Reserved.
+
+
+#include "Player/STUPlayerCharacter.h"
+#include "Camera/CameraComponent.h"
+#include "Components/InputComponent.h"
+#include "GameFramework/SpringArmComponent.h"
+#include "Components/STUWeaponComponent.h"
+
+ASTUPlayerCharacter::ASTUPlayerCharacter(const FObjectInitializer& ObjInt)
+    : Super(ObjInt)
+{
+    PrimaryActorTick.bCanEverTick = true;
+
+    SpringArmComponent = CreateDefaultSubobject<USpringArmComponent>("SpringArmComponent");
+    SpringArmComponent->SetupAttachment(GetRootComponent());
+    SpringArmComponent->bUsePawnControlRotation = true;
+    SpringArmComponent->SocketOffset = FVector(0.0f, 100.0f, 80.0f);
+
+    CameraComponent = CreateDefaultSubobject<UCameraComponent>("CameraComponent");
+    CameraComponent->SetupAttachment(SpringArmComponent);
+}
+
+void ASTUPlayerCharacter::SetupPlayerInputComponent(UInputComponent* PlayerInputComponent)
+{
+    Super::SetupPlayerInputComponent(PlayerInputComponent);
+    check(WeaponComponent);
+
+    PlayerInputComponent->BindAxis("MoveForward", this, &ASTUPlayerCharacter::MoveForwarder);
+    PlayerInputComponent->BindAxis("MoveRight", this, &ASTUPlayerCharacter::MoveRighter);
+    PlayerInputComponent->BindAxis("LookUp", this, &ASTUPlayerCharacter::AddControllerPitchInput);
+    PlayerInputComponent->BindAxis("TurnAround", this, &ASTUPlayerCharacter::AddControllerYawInput);
+    PlayerInputComponent->BindAction("Jump", IE_Pressed, this, &ASTUPlayerCharacter::Jump);
+    PlayerInputComponent->BindAction("Run", IE_Pressed, this, &ASTUPlayerCharacter::OnStartRun);
+    PlayerInputComponent->BindAction("Run", IE_Released, this, &ASTUPlayerCharacter::OnStopRun);
+    PlayerInputComponent->BindAction("Fire", IE_Pressed, WeaponComponent, &USTUWeaponComponent::StartFire);
+    PlayerInputComponent->BindAction("Fire", IE_Released, WeaponComponent, &USTUWeaponComponent::StopFire);
+    PlayerInputComponent->BindAction("NextWeapon", IE_Released, WeaponComponent, &USTUWeaponComponent::NextWeapon);
+    PlayerInputComponent->BindAction("Reload", IE_Pressed, WeaponComponent, &USTUWeaponComponent::Reload);
+}
+
+bool ASTUPlayerCharacter::IsRunn() const
+{
+    return WantsToRun && IsMovingForward && !GetVelocity().IsZero();
+}
+
+void ASTUPlayerCharacter::MoveForwarder(float Amount)
+{
+    IsMovingForward = Amount > 0.0f;
+    if (Amount == 0.0f)
+        return;
+
+    AddMovementInput(GetActorForwardVector(), Amount);
+}
+
+void ASTUPlayerCharacter::MoveRighter(float Amount)
+{
+    if (Amount == 0.0f)
+        return;
+
+    AddMovementInput(GetActorRightVector(), Amount);
+}
+
+void ASTUPlayerCharacter::OnStartRun()
+{
+    WantsToRun = true;
+}
+
+void ASTUPlayerCharacter::OnStopRun()
+{
+    WantsToRun = false;
+}
+
+void ASTUPlayerCharacter::OnDeath()
+{
+    Super::OnDeath();
+    
+    if (Controller)
+    {
+        Controller->ChangeState(NAME_Spectating);
+    }
+}
