+++ Source/MyShooterThemUp/Public/Player/STUBaseCharacter.h
@@ -6,10 +6,7 @@
 #include "GameFramework/Character.h"
 #include "STUBaseCharacter.generated.h"
 
-class UCameraComponent;
-class USpringArmComponent;
 class USTUHealthComponent;
-class UTextRenderComponent;
 class USTUWeaponComponent;
 
 UCLASS()
@@ -18,22 +15,13 @@ class MYSHOOTERTHEMUP_API ASTUBaseCharacter : public ACharacter
     GENERATED_BODY()
 
 public:
-    // Sets default values for this character's properties
     ASTUBaseCharacter(const FObjectInitializer& ObjInt);
 
 protected:
-    UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
-    USpringArmComponent* SpringArmComponent;
-
-    UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
-    UCameraComponent* CameraComponent;
 
     UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
     USTUHealthComponent* HealthComponent;
-
-    UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
-    UTextRenderComponent* HealthTextComponent;
-
+    
     UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
     USTUWeaponComponent* WeaponComponent;
 
@@ -53,11 +41,9 @@ protected:
 
 public:
     virtual void Tick(float DeltaTime) override;
-
-    virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;
-
+    
     UFUNCTION(BlueprintCallable, category = "Movement")
-    bool IsRunn() const;
+    virtual bool IsRunn() const;
 
     UFUNCTION(BlueprintCallable, category = "Movement")
     float GetMovementDirection() const;
@@ -67,14 +53,6 @@ public:
     void SetPlayerColor(const FLinearColor& Color);
 
 private:
-    bool WantsToRun = false;
-    bool IsMovingForward = false;
-
-    void MoveForwarder(float Amount);
-    void MoveRighter(float Amount);
-
-    void OnStartRun();
-    void OnStopRun();
     void OnHealthChanged(float Health, float HealthDelta);
 
     UFUNCTION()
