+++ Source/MyShooterThemUp/Public/Player/STUPlayerCharacter.h
@@ -1,0 +1,43 @@
+// Shoot Them Up Game, All Rights Reserved.
+
+#pragma once
+
+#include "CoreMinimal.h"
+#include "Player/STUBaseCharacter.h"
+#include "STUPlayerCharacter.generated.h"
+
+class UCameraComponent;
+class USpringArmComponent;
+
+UCLASS()
+class MYSHOOTERTHEMUP_API ASTUPlayerCharacter : public ASTUBaseCharacter
+{
+	GENERATED_BODY()
+
+public:
+    ASTUPlayerCharacter(const FObjectInitializer& ObjInt);
+
+protected:
+    UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
+    USpringArmComponent* SpringArmComponent;
+
+    UPROPERTY(VisibleAnywhere, BlueprintReadWrite, Category = "Componets")
+    UCameraComponent* CameraComponent;
+
+    virtual void OnDeath() override;
+
+public:
+    virtual void SetupPlayerInputComponent(class UInputComponent* PlayerInputComponent) override;
+
+    virtual bool IsRunn() const override;
+
+private:
+    bool WantsToRun = false;
+    bool IsMovingForward = false;
+
+    void MoveForwarder(float Amount);
+    void MoveRighter(float Amount);
+
+    void OnStartRun();
+    void OnStopRun();
+};
